backend/requirements.txt
pymongo 
langchain 
google-generativeai 
faker 
fastapi 
uvicorn
python-dotenv
langchain-google-genai
